services/transcription/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
openai-whisper==20231117
openai>=1.0.0  
torch==2.1.1
python-dotenv==1.0.0
pydantic==2.5.2